git-svn-id: http://192.168.1.50/svn/bhp_content_type_map@4 179f8c61-186b-4a6f-a3fc-937048475198

+++ managers/content_type_map_manager.py
@@ -1,7 +1,7 @@
 from django.db import models
-from django.db.models import F
+from django.db.models import F, get_model
 from django.contrib.contenttypes.models import ContentType
-
+from django.db.utils import IntegrityError
 
 class ContentTypeMapManager(models.Manager):
 
@@ -15,11 +15,13 @@ class ContentTypeMapManager(models.Manager):
         """
         content_type_maps = super(ContentTypeMapManager, self).exclude(name = F('content_type__name'))
         for content_type_map in content_type_maps:
-            try:
-                content_type = ContentType.objects.get(name = content_type_map.name)
+
+            if get_model(content_type_map.app_label, content_type_map.model):
+                name = get_model(content_type_map.app_label, content_type_map.model)._meta.module_name
+                content_type = ContentType.objects.get(model = name)
                 content_type_map.content_type = content_type
                 content_type_map.save()
-            except:    
+            else:    
                 content_type_map.delete()
     
     def populate(self):
@@ -29,10 +31,30 @@ class ContentTypeMapManager(models.Manager):
         content_types = ContentType.objects.all()        
         for content_type in content_types:
             if not super(ContentTypeMapManager, self).filter(content_type = content_type):
-                super(ContentTypeMapManager, self).create(
-                    content_type = content_type,
-                    app_label = content_type.app_label,
-                    name = content_type.name,
-                    model = content_type.model,
-                    )
+                #if super(ContentTypeMapManager, self).filter(content_type__name = content_type.name):
+                #    super(ContentTypeMapManager, self).filter(content_type__name = content_type.name).delete()
 
+                if get_model(content_type.app_label, content_type.model):
+                    verbose_name = get_model(content_type.app_label, content_type.model)._meta.verbose_name
+                else:
+                    verbose_name = content_type.name                    
+
+                try:
+                    super(ContentTypeMapManager, self).create(
+                        content_type = content_type,
+                        app_label = content_type.app_label,
+                        name = verbose_name,
+                        model = content_type.model,
+                        )
+                except IntegrityError:   
+                    if verbose_name <> 'content type map' and verbose_name <> 'identifier tracker':   
+                        raise TypeError(verbose_name)
+                    pass                  
+        
+        #content_type_maps = super(ContentTypeMapManager, self).__dict__['model'].objects.all()        
+        #for content_type_map in content_type_maps:
+        #    if get_model(content_type_map.app_label, content_type_map.model):
+        #        verbose_name = get_model(content_type_map.app_label, content_type_map.model)._meta.verbose_name
+        #        if not content_type_map.name == verbose_name:
+        #            content_type_map.name = verbose_name
+        #            content_type_map.save()
